requirements.txt
open3d
ros2-numpy
gradio